+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * SEARCH COMMA selector
 * Fixed bug with ORDER BY 1,2,3
 * Added subqueries for INSERT/DELETE/UPDATE
+* First ECHO plugin released (REUQUIRE ECHO)
 
 ### 0.1.5 "San Marino" (12.05.2015 - 12.05.2015)
 * Added Meteor package (agershun:alasql) - still does not work - skeleton